Add Load (world_content) On branch load Your branch is up to date with 'origin/load'.
Changes to be committed:
	modified:   .gitignore
	new file:   ingestion.py
	modified:   main.py
	new file:   sql/copy_from_table.sql
	new file:   sql/drop_tables.sql
	new file:   sql/tables.sql
