+++ .gitignore
@@ -1,6 +1,6 @@
 # Byte-compiled / optimized / DLL files
 .idea/
-
+csv/
 __pycache__/
 *.py[codz]
 *$py.class
