+++ sql/drop_tables.sql
@@ -1,0 +1,87 @@
+drop table if exists DestinyDamageTypeDefinition;
+drop table if exists DestinyArtifactDefinition;
+drop table if exists DestinyTraitDefinition;
+drop table if exists DestinySocialCommendationNodeDefinition;
+drop table if exists DestinyFireteamFinderLabelGroupDefinition;
+drop table if exists DestinyCharacterCustomizationCategoryDefinition;
+drop table if exists DestinyPlaceDefinition;
+drop table if exists DestinyInventoryItemLiteDefinition;
+drop table if exists DestinyGuardianRankConstantsDefinition;
+drop table if exists DestinyFireteamFinderLabelDefinition;
+drop table if exists DestinyInventoryBucketDefinition;
+drop table if exists DestinyRewardMappingDefinition;
+drop table if exists DestinyInventoryItemDefinition;
+drop table if exists DestinyItemTierTypeDefinition;
+drop table if exists DestinyLoadoutIconDefinition;
+drop table if exists DestinyUnlockEventDefinition;
+drop table if exists DestinyActivityModeDefinition;
+drop table if exists DestinyBondDefinition;
+drop table if exists DestinySackRewardItemListDefinition;
+drop table if exists DestinyEquipmentSlotDefinition;
+drop table if exists DestinyCharacterCustomizationOptionDefinition;
+drop table if exists DestinyVendorGroupDefinition;
+drop table if exists DestinyProgressionLevelRequirementDefinition;
+drop table if exists DestinyAchievementDefinition;
+drop table if exists DestinyLoreDefinition;
+drop table if exists DestinyUnlockExpressionMappingDefinition;
+drop table if exists DestinySocialCommendationDefinition;
+drop table if exists DestinyFactionDefinition;
+drop table if exists DestinyObjectiveDefinition;
+drop table if exists DestinyStatGroupDefinition;
+drop table if exists DestinyEntitlementOfferDefinition;
+drop table if exists DestinyPlatformBucketMappingDefinition;
+drop table if exists DestinyPresentationNodeDefinition;
+drop table if exists DestinyFireteamFinderActivityGraphDefinition;
+drop table if exists DestinyPlugSetDefinition;
+drop table if exists DestinyFireteamFinderConstantsDefinition;
+drop table if exists DestinyRewardSheetDefinition;
+drop table if exists DestinyLocationDefinition;
+drop table if exists DestinySandboxPerkDefinition;
+drop table if exists DestinyPowerCapDefinition;
+drop table if exists DestinyMaterialRequirementSetDefinition;
+drop table if exists DestinyActivityTypeDefinition;
+drop table if exists DestinyUnlockCountMappingDefinition;
+drop table if exists DestinyClassDefinition;
+drop table if exists DestinyReportReasonCategoryDefinition;
+drop table if exists DestinyGenderDefinition;
+drop table if exists DestinySeasonDefinition;
+drop table if exists DestinyStatDefinition;
+drop table if exists DestinyRewardSourceDefinition;
+drop table if exists DestinyActivityGraphDefinition;
+drop table if exists DestinyItemCategoryDefinition;
+drop table if exists DestinySandboxPatternDefinition;
+drop table if exists DestinyArtDyeChannelDefinition;
+drop table if exists DestinyLoadoutConstantsDefinition;
+drop table if exists DestinyFireteamFinderOptionDefinition;
+drop table if exists DestinyUnlockDefinition;
+drop table if exists DestinyProgressionMappingDefinition;
+drop table if exists DestinyMedalTierDefinition;
+drop table if exists DestinySocketCategoryDefinition;
+drop table if exists DestinyBreakerTypeDefinition;
+drop table if exists DestinyVendorDefinition;
+drop table if exists DestinyArtDyeReferenceDefinition;
+drop table if exists DestinyMilestoneDefinition;
+drop table if exists DestinyRecordDefinition;
+drop table if exists DestinyLoadoutColorDefinition;
+drop table if exists DestinyFireteamFinderActivitySetDefinition;
+drop table if exists DestinyCollectibleDefinition;
+drop table if exists DestinyUnlockValueDefinition;
+drop table if exists DestinyActivityInteractableDefinition;
+drop table if exists DestinyActivityDefinition;
+drop table if exists DestinyGlobalConstantsDefinition;
+drop table if exists DestinyDestinationDefinition;
+drop table if exists DestinyLoadoutNameDefinition;
+drop table if exists DestinyRewardItemListDefinition;
+drop table if exists DestinyRewardAdjusterProgressionMapDefinition;
+drop table if exists DestinyRewardAdjusterPointerDefinition;
+drop table if exists DestinyEnergyTypeDefinition;
+drop table if exists DestinyMetricDefinition;
+drop table if exists DestinyActivityModifierDefinition;
+drop table if exists DestinyRaceDefinition;
+drop table if exists DestinySeasonPassDefinition;
+drop table if exists DestinyEventCardDefinition;
+drop table if exists DestinyGuardianRankDefinition;
+drop table if exists DestinyChecklistDefinition;
+drop table if exists DestinyProgressionDefinition;
+drop table if exists DestinyFireteamFinderOptionGroupDefinition;
+drop table if exists DestinySocketTypeDefinition;
